+++ lib-common/blob-utf8.c
@@ -114,7 +114,7 @@ static const char * const __cp1252_or_latin9_to_utf8[0x40] = {
 };
 
 /* OG: this function should be made faster for the ASCII subset */
-static ssize_t blob_latin1_to_utf8_aux(blob_t *out, const char *s, int len,
+static int blob_latin1_to_utf8_aux(blob_t *out, const char *s, int len,
                                        int limit)
 {
     int res = 0;
@@ -140,12 +140,12 @@ static ssize_t blob_latin1_to_utf8_aux(blob_t *out, const char *s, int len,
     return res;
 }
 
-ssize_t blob_latin1_to_utf8(blob_t *out, const char *s, int len)
+int blob_latin1_to_utf8(blob_t *out, const char *s, int len)
 {
     return blob_latin1_to_utf8_aux(out, s, len, 0xa0);
 }
 
-ssize_t blob_latin9_to_utf8(blob_t *out, const char *s, int len)
+int blob_latin9_to_utf8(blob_t *out, const char *s, int len)
 {
     return blob_latin1_to_utf8_aux(out, s, len, 0xc0);
 }
